Add benchmark executor
